+++ packages/amino/src/coins.spec.ts
@@ -2,7 +2,7 @@ import { coin, coins, parseCoins } from "./coins";
 
 describe("coins", () => {
   describe("coin", () => {
-    it("works for basic values", () => {
+    it("works for number amounts", () => {
       expect(coin(123, "utoken")).toEqual({ amount: "123", denom: "utoken" });
       expect(coin(123.0, "utoken")).toEqual({ amount: "123", denom: "utoken" });
       expect(coin(Number.MAX_SAFE_INTEGER, "utoken")).toEqual({
@@ -14,22 +14,56 @@ describe("coins", () => {
     });
 
     it("throws for non-safe-integer values", () => {
-      expect(() => coin(1.23, "utoken")).toThrow();
-      expect(() => coin(NaN, "utoken")).toThrow();
-      expect(() => coin(Number.POSITIVE_INFINITY, "utoken")).toThrow();
-      expect(() => coin(Number.MAX_SAFE_INTEGER + 1, "utoken")).toThrow();
+      expect(() => coin(1.23, "utoken")).toThrowError(/Given amount is not a safe integer/i);
+      expect(() => coin(NaN, "utoken")).toThrowError(/Given amount is not a safe integer/i);
+      expect(() => coin(Number.POSITIVE_INFINITY, "utoken")).toThrowError(
+        /Given amount is not a safe integer/i,
+      );
+      expect(() => coin(Number.MAX_SAFE_INTEGER + 1, "utoken")).toThrowError(
+        /Given amount is not a safe integer/i,
+      );
     });
 
     it("throws for negative values", () => {
-      expect(() => coin(-1, "utoken")).toThrow();
-      expect(() => coin(Number.MIN_SAFE_INTEGER, "utoken")).toThrow();
-      expect(() => coin(Number.NEGATIVE_INFINITY, "utoken")).toThrow();
+      expect(() => coin(-1, "utoken")).toThrowError(/Given amount is not a safe integer/i);
+      expect(() => coin(Number.MIN_SAFE_INTEGER, "utoken")).toThrowError(
+        /Given amount is not a safe integer/i,
+      );
+      expect(() => coin(Number.NEGATIVE_INFINITY, "utoken")).toThrowError(
+        /Given amount is not a safe integer/i,
+      );
+    });
+
+    it("works for string amounts", () => {
+      expect(coin("0", "utoken")).toEqual({ amount: "0", denom: "utoken" });
+      expect(coin("1", "utoken")).toEqual({ amount: "1", denom: "utoken" });
+      expect(coin("00123", "utoken")).toEqual({ amount: "123", denom: "utoken" });
+      expect(coin("12300", "utoken")).toEqual({ amount: "12300", denom: "utoken" });
+      expect(coin("9007199254740992", "utoken")).toEqual({ amount: "9007199254740992", denom: "utoken" });
+      // ETH supply (~118 mio ETH)
+      expect(coin("118273505060000000000000000", "wei")).toEqual({
+        amount: "118273505060000000000000000",
+        denom: "wei",
+      });
+    });
+
+    it("throws for invalid amount strings", () => {
+      expect(() => coin("-1", "utoken")).toThrowError(/Invalid unsigned integer string format/i);
+      expect(() => coin("0x01", "utoken")).toThrowError(/Invalid unsigned integer string format/i);
+      expect(() => coin("NaN", "utoken")).toThrowError(/Invalid unsigned integer string format/i);
+      expect(() => coin("1.0", "utoken")).toThrowError(/Invalid unsigned integer string format/i);
+      expect(() => coin("1 ", "utoken")).toThrowError(/Invalid unsigned integer string format/i);
+      expect(() => coin(" 1", "utoken")).toThrowError(/Invalid unsigned integer string format/i);
+      expect(() => coin("1.1827350506e+26", "utoken")).toThrowError(
+        /Invalid unsigned integer string format/i,
+      );
     });
   });
 
   describe("coins", () => {
     it("returns one element array of coin", () => {
       expect(coins(123, "utoken")).toEqual([{ amount: "123", denom: "utoken" }]);
+      expect(coins("123", "utoken")).toEqual([{ amount: "123", denom: "utoken" }]);
     });
   });