+++ packages/amino/src/coins.ts
@@ -7,15 +7,39 @@ export interface Coin {
 
 /**
  * Creates a coin.
+ *
+ * If your values do not exceed the safe integer range of JS numbers (53 bit),
+ * you can use the number type here. This is the case for all typical Cosmos SDK
+ * chains that use the default 6 decimals.
+ *
+ * In case you need to supportr larger values, use unsigned integer strings instead.
  */
-export function coin(amount: number, denom: string): Coin {
-  return { amount: new Uint53(amount).toString(), denom: denom };
+export function coin(amount: number | string, denom: string): Coin {
+  let outAmount: string;
+  if (typeof amount === "number") {
+    try {
+      outAmount = new Uint53(amount).toString();
+    } catch (_err) {
+      throw new Error(
+        "Given amount is not a safe integer. Consider using a string instead to overcome the limitations of JS numbers.",
+      );
+    }
+  } else {
+    if (!amount.match(/^[0-9]+$/)) {
+      throw new Error("Invalid unsigned integer string format");
+    }
+    outAmount = amount.replace(/^0*/, "") || "0";
+  }
+  return {
+    amount: outAmount,
+    denom: denom,
+  };
 }
 
 /**
  * Creates a list of coins with one element.
  */
-export function coins(amount: number, denom: string): Coin[] {
+export function coins(amount: number | string, denom: string): Coin[] {
   return [coin(amount, denom)];
 }